refactor: Moved app/data to lib/data and change api/data to api/routes
Moved app/data to lib/data and change api/data to api/routes
