+++ app/api/routes/chord/route.ts
@@ -1,11 +1,11 @@
 /**
  * @title Chord Diagram Data API
  * @fileoverview Chord diagram data API
- * @path /app/api/data/chord/route.ts
+ * @path /app/api/routes/chord/route.ts
  */
 
 import { NextResponse } from 'next/server';
-import { geneExpressionChordData } from '@/app/data/chord_data';
+import { geneExpressionChordData } from '@/lib/data/chord_data';
 
 // Chord diagram data API
 export async function GET() {
