+++ app/api/routes/differential-expression/route.ts
@@ -1,11 +1,11 @@
 /**
  * @title Differential Expression Data API
  * @fileoverview Differential expression data API
- * @path /app/api/data/differential-expression/route.ts
+ * @path /app/api/routes/differential-expression/route.ts
  */
 
 import { NextResponse } from 'next/server';
-import { differentialExpressionMockData } from '@/app/data/differential_expression_data';
+import { differentialExpressionMockData } from '@/lib/data/differential_expression_data';
 
 export async function GET() {
   // Return the mock data for differential expression
