+++ components/visualizations/DegRankings.tsx
@@ -32,7 +32,7 @@ export const DegRankings: React.FC = () => {
 
   const fetchRankingsData = async () => {
     try {
-      const response = await fetch('/api/data/deg-rankings');
+      const response = await fetch('/api/routes/deg-rankings');
       if (!response.ok) {
         throw new Error(`HTTP error! status: ${response.status}`);
       }
